+++ erraux/doc.go
@@ -1,0 +1,5 @@
+/*
+Package erraux provides auxiliary functionality for marshaling and returning
+errors over HTTP.
+*/
+package erraux